requirements.txt
aiofiles==0.8.0
discord.py==1.7.3
jishaku==2.5.1
python-dotenv==0.20.0
